Extract MarketViewItem for debugging
Debugging freezes app. Extracting MarketViewItem to separate class helped.
